+++ examples/simple/settings/base.py
@@ -322,6 +322,7 @@ INSTALLED_APPS = [
     # ************************ Security elements ****************************
     # ***********************************************************************
     'fobi.contrib.plugins.form_elements.security.honeypot',
+    'fobi.contrib.plugins.form_elements.security.invisible_recaptcha',
 
     # ***********************************************************************
     # ************************* Testing elements ****************************
@@ -378,6 +379,10 @@ INSTALLED_APPS = [
     # CKEditor widget
     'fobi.contrib.themes.bootstrap3.widgets.form_elements.'
     'content_richtext_bootstrap3_widget',
+
+    # Invisible Recaptcha widget
+    'fobi.contrib.themes.bootstrap3.widgets.form_elements.'
+    'invisible_recaptcha_bootstrap3_widget',
 
     # # Markdown
     # 'fobi.contrib.themes.bootstrap3.widgets.form_elements.'