+++ src/fobi/contrib/themes/bootstrap3/widgets/form_elements/invisible_recaptcha_bootstrap3_widget/__init__.py
@@ -1,0 +1,9 @@
+__title__ = 'fobi.contrib.themes.bootstrap3.widgets.form_elements.' \
+            'invisible_recaptcha_bootstrap3_widget'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = '2014-2018 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = ('default_app_config',)
+
+default_app_config = 'fobi.contrib.themes.bootstrap3.widgets.form_elements.' \
+                     'invisible_recaptcha_bootstrap3_widget.apps.Config'